feat: implemented xhr queue
