+++ packages/analytics-js-plugins/src/dataplaneEventsQueue/index.ts
@@ -3,24 +3,14 @@ import {
   ExtensionPlugin,
   PluginName,
   ApplicationState,
-  QueueOpts,
   RudderEvent,
   ILogger,
-  IErrorHandler,
-  IHttpClient,
-  HttpClient,
-  DoneCallback
+  IPluginsManager
 } from '../types/common';
-import { getDeliveryPayload, validatePayloadSize, getDeliveryUrl, getNormalizedQueueOptions } from './utilities';
-import { QUEUE_NAME, REQUEST_TIMEOUT_MS } from './constants';
-import { Queue, getCurrentTimeFormatted, toBase64 } from '../utilities/common';
-import { XHRQueueItem } from './types';
 
 const pluginName = PluginName.DataplaneEventsQueue;
 
-let eventsQueue: Queue;
-let dataplaneUrl: string;
-let httpClient: IHttpClient;
+let pluginsManager: IPluginsManager;
 let initialized = false;
 
 const DataplaneEventsQueue = (): ExtensionPlugin => ({
@@ -30,58 +20,15 @@ const DataplaneEventsQueue = (): ExtensionPlugin => ({
     state.plugins.loadedPlugins.value = [...state.plugins.loadedPlugins.value, pluginName];
   },
   dataplaneEventsQueue: {
-    init(writeKey: string, inDataplaneUrl: string, queueOpts: QueueOpts, errorHandler?: IErrorHandler, logger?: ILogger): void {
+    init(inPluginsManager: IPluginsManager, state: ApplicationState): void {
       if (initialized) {
         return;
       }
-      dataplaneUrl = inDataplaneUrl;
 
-      httpClient = new HttpClient(errorHandler, logger);
-      httpClient.setAuthHeader(writeKey);
+      pluginsManager = inPluginsManager;
 
-      const finalQOpts = getNormalizedQueueOptions(queueOpts);
-
-      eventsQueue = new Queue(QUEUE_NAME, finalQOpts, (item: XHRQueueItem, done: DoneCallback, willBeRetried: boolean) => {
-        const { url, event, headers } = item;
-        // Update sentAt timestamp to the latest timestamp
-        event.sentAt = getCurrentTimeFormatted();
-        const data = getDeliveryPayload(event);
-
-        if (data) {
-          httpClient.getAsyncData({
-            url,
-            options: {
-              method: 'POST',
-              headers,
-              data,
-            },
-            timeout: REQUEST_TIMEOUT_MS,
-            callback: result => {
-              if (result !== undefined) {
-                let errMsg = `Unable to deliver event to ${url}.`;
-                
-                if (willBeRetried) {
-                  errMsg = `${errMsg} It'll be retried.`;
-                }
-                else {
-                  errMsg = `${errMsg} It'll be dropped.`;
-                }
-                logger?.error(errMsg);
-
-                // failed
-                done(result);
-              } else {
-                // success
-                done(null);
-              }
-            },
-          });
-        } else {
-          logger?.error(`Unable to prepare event payload for delivery. It'll be dropped.`);
-          // Mark the item as done so that it can be removed from the queue
-          done(null);
-        }
-      });
+      // TODO: Based on state initialize only the required plugin
+      pluginsManager.invokeSingle('xhrDeliveryQueue', 'init', state);
 
       initialized = true;
     },
@@ -94,7 +41,8 @@ const DataplaneEventsQueue = (): ExtensionPlugin => ({
       if (!initialized) {
         return;
       }
-      eventsQueue.start();
+
+      pluginsManager.invokeSingle('xhrDeliveryQueue', 'start');
     },
 
     /**
@@ -108,23 +56,7 @@ const DataplaneEventsQueue = (): ExtensionPlugin => ({
         return;
       }
 
-      // sentAt is only added here for the validation step
-      // It'll be updated to the latest timestamp during actual delivery
-      event.sentAt = getCurrentTimeFormatted();
-      validatePayloadSize(event, logger);
-
-      const url = getDeliveryUrl(dataplaneUrl, event.type);
-      // Other default headers are added by the HttpClient
-      // Auth header is added during initialization
-      const headers = {
-        AnonymousId: toBase64(event.anonymousId),
-      };
-
-      eventsQueue.addItem({
-        url,
-        headers,
-        event,
-      } as XHRQueueItem);
+      pluginsManager.invokeSingle('xhrDeliveryQueue', 'enqueue', event, logger);
     },
   },
 });
