+++ packages/analytics-js-plugins/src/utilities/common.ts
@@ -1,3 +1,5 @@
 export { Queue } from "@rudderstack/analytics-js/npmPackages/localstorage-retry";
 export { getCurrentTimeFormatted } from "@rudderstack/analytics-js/components/utilities/timestamp";
 export { toBase64 } from "@rudderstack/analytics-js/components/utilities/string";
+export { mergeDeepRight } from "@rudderstack/analytics-js/components/utilities/object";
+export { replaceNullValues } from '@rudderstack/analytics-js/components/utilities/json';
