+++ packages/analytics-js-plugins/src/xhrQueue/constants.ts
