+++ packages/analytics-js-plugins/src/xhrQueue/types.ts
